requirements.txt
black==24.4.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
flake8==7.1.0
fonttools==4.53.0
kiwisolver==1.4.5
llvmlite==0.43.0
matplotlib==3.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==2.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.14.0
six==1.16.0
